fix two -Wextra warnings

+++ debsig-verify.c
@@ -392,7 +392,8 @@ int main(int argc, char *argv[]) {
     while ((pd_ent = readdir(pd)) != NULL && (pol == NULL || list_only)) {
 	char *ext = strstr(pd_ent->d_name, ".pol");
 	/* Make sure we have the right name format */
-	if (ext == NULL || (ext - pd_ent->d_name) + 4 != strlen(pd_ent->d_name))
+	if (ext == NULL || 
+            (size_t)(ext - pd_ent->d_name) + 4 != strlen(pd_ent->d_name))
 	    continue;
 
 	if (force_file != NULL && strcmp(pd_ent->d_name, force_file))
